requirements.txt
gensim==4.3.1
pinecone-client==2.2.1
numpy==1.25.0
tqdm==4.65.0
python-dotenv==1.0.0
fastapi
uvicorn[standard]

